+++ controls/tests.py
@@ -15,7 +15,7 @@ from pathlib import PurePath
 import tempfile
 from django.test import TestCase
 from django.utils.text import slugify
-from selenium.common.exceptions import WebDriverException
+from selenium.common.exceptions import WebDriverException, NoSuchElementException, StaleElementReferenceException
 from selenium.webdriver.common.by import By
 from selenium.webdriver.support import expected_conditions as EC
 from selenium.webdriver.support.select import Select
@@ -35,6 +35,7 @@ from system_settings.models import SystemSettings
 
 from urllib.parse import urlparse
 
+from guardian.shortcuts import assign_perm
 
 #####################################################################
 
@@ -547,6 +548,153 @@ class ElementUnitTests(TestCase):
         self.assertTrue(e2.id is not None)
         self.assertTrue(e2.component_type == "hardware")
         self.assertTrue(e2.component_state == "disposition")
+
+class ElementUITests(OrganizationSiteFunctionalTests):
+
+    def test_element_create_form(self):
+        self.browser.get(self.url("/"))
+        self._login()
+        self.browser.get(self.url("/controls/components"))
+        self.click_element("#component-new")
+        wait_for_sleep_after(lambda: self.assertInNodeText("New Component", "h2"))
+        element_name = "Tanium"
+        self.clear_and_fill_field("#id_name", element_name)
+        self.clear_and_fill_field("#id_full_name", "Tanium Software")
+        self.clear_and_fill_field("#id_description", "Tanium")
+        # Submit form
+        self.click_element("#create-element-button")
+        # Check element/component page
+        wait_for_sleep_after(lambda: self.assertInNodeText(element_name, "#element-name"))
+        var_sleep(1)
+        lockIsPresent =  wait_for_sleep_after(lambda: self.browser.find_element_by_id('lock-glyph'))
+        self.assertTrue(lockIsPresent)
+        var_sleep(1)
+
+    def test_edit_element_private_to_public(self):
+        self.browser.get(self.url("/"))
+        self._login()
+        self.browser.get(self.url("/controls/components"))
+        self.click_element("#component-new")
+        wait_for_sleep_after(lambda: self.assertInNodeText("New Component", "h2"))
+        element_name = "Tanium"
+        self.clear_and_fill_field("#id_name", element_name)
+        self.clear_and_fill_field("#id_full_name", "Tanium Software")
+        self.clear_and_fill_field("#id_description", "Tanium")
+        var_sleep(1)
+        # Submit form
+        self.click_element("#create-element-button")
+        # Check element/component page
+        wait_for_sleep_after(lambda: self.assertInNodeText(element_name, "#element-name"))
+        var_sleep(1)
+
+        # Check if lock glyphicon is present on the document to signify the Element is private
+        lockIsPresent =  wait_for_sleep_after(lambda: self.browser.find_element_by_id('lock-glyph'))
+        self.assertTrue(lockIsPresent)
+
+        # Edit component    
+        self.click_element("#edit-button")
+        # Check if at Edit Component modal
+        wait_for_sleep_after(lambda: self.assertInNodeText("Edit Component", "#invitation_modal_title"))
+        var_sleep(1)
+        
+        # Click on checkbox and save
+        self.click_element("#private-input")
+        var_sleep(1)
+
+        # Click Submit button
+        self.click_element("#submit-btn")
+
+        # Check if lock glyph is displayed 
+        notThere = False
+        # Expecting an exception to be thrown due to the element not being on the page
+        try:
+            wait_for_sleep_after(lambda: self.browser.find_element_by_id('lock-glyph'))
+        except NoSuchElementException:
+            print('Lock Glyph does not exist on page')
+            pass
+        else:
+            notThere = True
+            pass
+        self.assertTrue(notThere)
+        var_sleep(3)
+
+        # Changing state of Element back to private
+        self.click_element("#edit-button")
+        # Check if at Edit Component modal
+        wait_for_sleep_after(lambda: self.assertInNodeText("Edit Component", "#invitation_modal_title"))
+        var_sleep(1)
+
+        # Click on checkbox and save
+        self.click_element("#private-input")
+        var_sleep(1)
+
+        # Click submit
+        self.click_element("#submit-btn")
+        var_sleep(3)
+
+        # Check if lock glyph is displayed 
+        glyphPresent = wait_for_sleep_after(lambda: self.browser.find_element_by_id('lock-glyph'))
+        
+        self.assertTrue(glyphPresent)
+        var_sleep(1)
+
+    def test_if_viewer_can_edit_element(self):
+        self.browser.get(self.url("/"))
+        self._login()
+        self.browser.get(self.url("/controls/components"))
+        self.click_element("#component-new")
+        wait_for_sleep_after(lambda: self.assertInNodeText("New Component", "h2"))
+        element_name = "Tanium"
+        self.clear_and_fill_field("#id_name", element_name)
+        self.clear_and_fill_field("#id_full_name", "Tanium Software")
+        self.clear_and_fill_field("#id_description", "Tanium")
+        # Submit form
+        self.click_element("#create-element-button")
+        # Check element/component page
+        wait_for_sleep_after(lambda: self.assertInNodeText(element_name, "#element-name"))
+        var_sleep(1)
+        lockIsPresent =  wait_for_sleep_after(lambda: self.browser.find_element_by_id('lock-glyph'))
+        self.assertTrue(lockIsPresent)
+        var_sleep(3)
+
+        # Assign view_element role to another user
+        print('ASSIGNING VIEW PERMISSION TO USER2\n')
+        currentElement = Element.objects.get(name=element_name)
+        # currentElement.assign_owner_permissions(self.user)
+        assign_perm('view_element', self.user2, currentElement)
+        print('ASSIGNED USER2\n')
+
+        # Sign out -> Login with another user -> Check if we can edit
+        # Log out
+        self.browser.get(self.url("/accounts/logout/"))
+        var_sleep(1)
+
+        # Sign in as User2
+        self._login(self.user2.username, self.user2.clear_password)
+        var_sleep(2)
+
+        self.browser.get(self.url("/controls/components"))
+        var_sleep(1)
+
+        # Grab Tanium link element
+        wait_for_sleep_after(lambda: self.browser.find_element_by_link_text("Tanium").click())
+        var_sleep(1)
+
+        # Check element/component page
+        wait_for_sleep_after(lambda: self.assertInNodeText(element_name, "#element-name"))
+        var_sleep(1)
+
+        print('Checking if edit button is visible')
+        # Edit button should not be visible
+        notThere = False
+        try:
+            wait_for_sleep_after(lambda: self.browser.find_element_by_id('edit-button'))
+        except NoSuchElementException:
+            print('Edit Button does not exist on page')
+            notThere = True
+            pass
+        self.assertTrue(notThere)
+        var_sleep(3)
 
 class ElementControlUnitTests(SeleniumTest):